Rename the Logger module
