+++ compiler/hie-core/test/Demo.hs
@@ -17,7 +17,7 @@ import Development.IDE.Types.Location
 import Data.String
 import Development.IDE.Types.Diagnostics
 import Development.IDE.Types.Options
-import Development.IDE.Logger
+import Development.IDE.Types.Logger
 import qualified Data.Text.IO as T
 import Language.Haskell.LSP.Messages
 import Development.IDE.LSP.LanguageServer
