+++ daml-foundations/daml-ghc/ide/test/Development/IDE/State/API/Testing.hs
@@ -43,7 +43,7 @@ import qualified Development.IDE.Types.Diagnostics as D
 import qualified Development.IDE.Types.Location as D
 import DA.Service.Daml.Compiler.Impl.Scenario as SS
 import Development.IDE.State.Rules.Daml
-import qualified Development.IDE.Logger as Logger
+import qualified Development.IDE.Types.Logger as Logger
 import DA.Daml.GHC.Compiler.Options (defaultOptionsIO)
 import Development.IDE.State.Service.Daml(VirtualResource(..))
 import DA.Test.Util (standardizeQuotes)
